requirements.txt
spotipy==2.25.1
pandas
python-dotenv
pendulum
psycopg2-binary